move all the config stuff into a new config/ folder and make it a subpackage by adding a config/__init__.py file. This will make it easier to continue refactoring the config system later
