+++ manim/__init__.py
@@ -1,12 +1,8 @@
 #!/usr/bin/env python
 
 # Importing the config module should be the first thing we do, since other
-# modules depend on the global config dict for initialization.  Note that the
-# global config dict is called 'config', just like the module itself.  That's
-# why we import the module first with a different name, and then the dict.
-from . import config as _config
-
-from .config import config, file_writer_config, camera_config, tempconfig
+# modules depend on the global config dict for initialization.
+from .config import *
 
 from .constants import *
