+++ manim/animation/indication.py
@@ -18,8 +18,8 @@ __all__ = [
 
 import numpy as np
 
+from .. import config
 from ..constants import *
-from ..config import config
 from ..animation.animation import Animation
 from ..animation.movement import Homotopy
 from ..animation.composition import AnimationGroup
