+++ manim/config/default.cfg
