+++ manim/mobject/frame.py
@@ -6,8 +6,8 @@ __all__ = [
 ]
 
 
+from .. import config
 from ..constants import *
-from ..config import config
 from ..mobject.geometry import Rectangle
 from ..utils.config_ops import digest_config
