+++ manim/mobject/svg/tex_mobject.py
@@ -12,8 +12,8 @@ __all__ = [
 from functools import reduce
 import operator as op
 
+from ... import config
 from ...constants import *
-from ...config import config
 from ...mobject.geometry import Line
 from ...mobject.svg.svg_mobject import SVGMobject
 from ...mobject.svg.svg_mobject import VMobjectFromSVGPathstring
