+++ manim/scene/scene.py
@@ -13,13 +13,12 @@ import copy
 from tqdm import tqdm as ProgressDisplay
 import numpy as np
 
+from .. import camera_config, file_writer_config, logger
 from ..animation.animation import Animation
 from ..animation.transform import MoveToTarget, ApplyMethod
 from ..camera.camera import Camera
 from ..constants import *
-from ..config import camera_config, file_writer_config
 from ..container import Container
-from ..logger import logger
 from ..mobject.mobject import Mobject
 from ..scene.scene_file_writer import SceneFileWriter
 from ..utils.iterables import list_update
